requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
disqus-python==0.4.2
Django==5.0.6
django-appconf==1.0.6
django-compressor==4.4
django-contrib-comments==2.2.0
django-debug-toolbar==4.3.0
django-iipimage==1.0
django-pagination==1.0.7
django-reversion==5.0.12
django-tinymce==4.0.0
feedparser==6.0.11
filebrowser-safe==1.1.1
grappelli-safe==1.1.1
html5lib==1.1
idna==3.7
importlib==1.0.4
lxml==5.2.2
Mezzanine==6.0.0
nose==1.3.7
oauthlib==3.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pytz==2024.1
rcssmin==1.1.1
regex==2024.5.10
requests==2.31.0
requests-oauthlib==2.0.0
reversion==0.2
rjsmin==1.2.1
setuptools==69.5.1
sgmllib3k==1.0.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tinycss2==1.2.1
toml==0.10.2
tzlocal==5.2
urllib3==2.2.1
webencodings==0.5.1
Whoosh==2.7.4
